+++ CHANGELOG.md
@@ -16,4 +16,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [Issue #22](https://github.com/manheim/jenkinsfile-pipeline-library/issues/22): Update codecov bash script integration
 - [Issue #27](https://github.com/manheim/jenkinsfile-pipeline-library/issues/27): Control the node where the pipeline is run
 - [Issue #29](https://github.com/manheim/jenkinsfile-pipeline-library/issues/29): Allow access to Jenkins Credentials during BuildStage
+- [Issue #31](https://github.com/manheim/jenkinsfile-pipeline-library/issues/31): Allow DeployStage to be decorated by Plugins